convert observable-object.js -> observable-object.ts
